requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.2
django-cors-headers==3.13.0
django-heroku==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
gunicorn==20.1.0
install==1.3.5
Pillow==9.2.0
PyJWT==2.5.0
python-decouple==3.8
pytz==2022.4
sqlparse==0.4.3
whitenoise==6.2.0
